Extract LiveMode
